Fix CodeQL-flagged int64->int32 truncation on the chronic-board windows
Both CHRONIC_BOARD_CLOSE_WINDOW_DAYS and CHRONIC_BOARD_WINDOW_DAYS were
read as int64 (worker.EnvInt64) then narrowed with a bare int32(...) cast
before reaching pgtype.Interval.Days — an out-of-range value would wrap
into an arbitrary, possibly negative, interval on a destructive-close code
path instead of failing the run. Adds worker.EnvInt32, which parses at
ParseInt's 32-bit size so an out-of-range value is a parse error rather
than a silent truncation, and switches both call sites and
closeChronicBoards's signature to int32 throughout.

+++ cmd/close-chronic-boards/run.go
@@ -32,9 +32,9 @@ type chronicBoardsReport struct {
 // a dry run genuinely writes nothing. A board with `board == ""` is a boardless provider's own
 // health record (see ingest-board-health spec) and routes to the source-scoped close/count
 // instead of the board-scoped one, since it already stands for the provider's whole catalogue.
-func closeChronicBoards(ctx context.Context, q *db.Queries, closeWindowDays int64, maxBoards int32, apply bool) (chronicBoardsReport, error) {
+func closeChronicBoards(ctx context.Context, q *db.Queries, closeWindowDays int32, maxBoards int32, apply bool) (chronicBoardsReport, error) {
 	rows, err := q.ListChronicBoards(ctx, db.ListChronicBoardsParams{
-		AgeWindow: pgtype.Interval{Days: int32(closeWindowDays), Valid: true},
+		AgeWindow: pgtype.Interval{Days: closeWindowDays, Valid: true},
 		MaxBoards: maxBoards,
 	})
 	if err != nil {
